refactor(gateway): serve /v1/models from an endpoint module instead of the router manager
Listing models is not routing: it reads the worker registry and, for a
caller presenting its own provider credential, fans out to the external
upstreams. It lived on `RouterManager` only because that type happened
to implement every endpoint, which also made the manager the sole holder
of an `AuthConfig` and an HTTP client.

Move the logic, unchanged, into `endpoints::models::list_models`, which
derives the gateway's own credentials from the router config the way the
manager did at startup. `RouterManager` loses its client and auth fields
and the `get_models` implementation; the `RouterTrait` default for
`get_models` goes with it since nothing implemented it any more. The
short-circuit test that keeps tenant keys from being forwarded upstream
moves with the code, driven through explicit collaborators.
