+++ model_gateway/src/routers/router_manager.rs
@@ -4,18 +4,16 @@
 //! - Single Router Mode (enable_igw=false): Router owns workers directly
 //! - Multi-Router Mode (enable_igw=true): RouterManager coordinates everything
 
-use std::{collections::HashSet, sync::Arc};
+use std::sync::Arc;
 
 use async_trait::async_trait;
 use axum::{
     body::Body,
     extract::Request,
-    http::{header, HeaderMap, HeaderValue, StatusCode},
+    http::{HeaderMap, StatusCode},
     response::{IntoResponse, Response},
-    Json,
 };
 use dashmap::DashMap;
-use futures::future::select_all;
 use openai_protocol::{
     chat::ChatCompletionRequest,
     classify::ClassifyRequest,
@@ -24,8 +22,6 @@ use openai_protocol::{
     generate::GenerateRequest,
     interactions::InteractionsRequest,
     messages::CreateMessageRequest,
-    model_card::ModelCard,
-    models::ListModelsResponse,
     realtime_session::{
         RealtimeClientSecretCreateRequest, RealtimeSessionCreateRequest,
         RealtimeTranscriptionSessionCreateRequest,
@@ -35,15 +31,14 @@ use openai_protocol::{
     transcription::{AudioFile, TranscriptionRequest},
     UNKNOWN_MODEL_ID,
 };
-use serde_json::Value;
 use tracing::{debug, info, warn};
 
 use crate::{
     app_context::AppContext,
     config::RoutingMode,
-    middleware::{AuthConfig, TenantRequestMeta},
+    middleware::TenantRequestMeta,
     routers::{
-        common::{body_policy::REASON_MODEL_SELECTION, header_utils::apply_provider_headers},
+        common::body_policy::REASON_MODEL_SELECTION,
         error as route_error,
         factory::{router_ids, RouterId},
         BodyPolicy, RouterFactory, RouterTrait,
@@ -54,24 +49,15 @@ use crate::{
 
 pub struct RouterManager {
     worker_registry: Arc<WorkerRegistry>,
-    client: reqwest::Client,
-    /// Every credential that authenticates as *this gateway* (shared
-    /// `api_key` plus any per-tenant keys) — not just the shared key. Used
-    /// to keep `/v1/models`' BYOK short-circuit from mistaking a valid
-    /// tenant-scoped key for a foreign upstream-provider credential and
-    /// forwarding it externally.
-    gateway_auth: AuthConfig,
     routers: Arc<DashMap<RouterId, Arc<dyn RouterTrait>>>,
     default_router: Arc<std::sync::RwLock<Option<RouterId>>>,
     enable_igw: bool,
 }
 
 impl RouterManager {
-    pub fn new(worker_registry: Arc<WorkerRegistry>, client: reqwest::Client) -> Self {
+    pub fn new(worker_registry: Arc<WorkerRegistry>) -> Self {
         Self {
             worker_registry,
-            client,
-            gateway_auth: AuthConfig::new(None),
             routers: Arc::new(DashMap::new()),
             default_router: Arc::new(std::sync::RwLock::new(None)),
             enable_igw: false,
@@ -100,15 +86,8 @@ impl RouterManager {
         config: &ServerConfig,
         app_context: &Arc<AppContext>,
     ) -> Result<Arc<Self>, String> {
-        let mut manager = Self::new(
-            app_context.worker_registry.clone(),
-            app_context.client.clone(),
-        );
+        let mut manager = Self::new(app_context.worker_registry.clone());
         manager.enable_igw = config.router_config.enable_igw;
-        manager.gateway_auth = AuthConfig::with_tenant_keys(
-            config.router_config.api_key.clone(),
-            &config.router_config.tenant_api_keys,
-        );
         let manager = Arc::new(manager);
 
         if config.router_config.enable_igw {
@@ -367,110 +346,6 @@ impl RouterManager {
                     .and_then(|id| self.routers.get(id).map(|r| r.clone()))
             })
     }
-
-    /// Build a response from self-hosted registry models (excludes external workers).
-    fn registry_models_response(&self) -> Response {
-        let cards: Vec<_> = self
-            .worker_registry
-            .get_all()
-            .iter()
-            .filter(|w| !matches!(w.metadata().spec.runtime_type, RuntimeType::External))
-            .flat_map(|w| w.models())
-            .collect();
-        if cards.is_empty() {
-            (StatusCode::SERVICE_UNAVAILABLE, "No models available").into_response()
-        } else {
-            let resp = ListModelsResponse::from_model_cards(cards);
-            (StatusCode::OK, Json(resp)).into_response()
-        }
-    }
-
-    /// Fan out to all healthy external upstreams concurrently with the caller's
-    /// bearer token and return the first successful model inventory. Returns an
-    /// empty vec on total failure.
-    async fn fetch_upstream_models(&self, bearer_token: &str) -> Vec<ModelCard> {
-        let unique_urls: Vec<_> = self
-            .worker_registry
-            .get_workers_filtered(None, None, None, Some(RuntimeType::External), true)
-            .iter()
-            .map(|w| w.url().to_string())
-            .collect::<HashSet<_>>()
-            .into_iter()
-            .collect();
-
-        if unique_urls.is_empty() {
-            return Vec::new();
-        }
-
-        debug!(
-            "Trying {} upstream(s) for model discovery",
-            unique_urls.len()
-        );
-
-        let auth = match HeaderValue::from_str(&format!("Bearer {bearer_token}")) {
-            Ok(v) => Some(v),
-            Err(e) => {
-                warn!("Bearer token contains invalid header characters: {e}");
-                return Vec::new();
-            }
-        };
-
-        // Fan out concurrently; return first non-empty result.
-        let mut pending: Vec<_> = unique_urls
-            .into_iter()
-            .map(|url| {
-                Box::pin(Self::fetch_models_from(
-                    self.client.clone(),
-                    url,
-                    auth.clone(),
-                ))
-            })
-            .collect();
-
-        while !pending.is_empty() {
-            let (cards, _index, remaining) = select_all(pending).await;
-            if !cards.is_empty() {
-                return cards;
-            }
-            pending = remaining;
-        }
-
-        Vec::new()
-    }
-
-    /// Fetch models from a single upstream endpoint.
-    async fn fetch_models_from(
-        client: reqwest::Client,
-        base_url: String,
-        auth: Option<HeaderValue>,
-    ) -> Vec<ModelCard> {
-        let url = format!("{base_url}/v1/models");
-        let req = apply_provider_headers(client.get(&url), &url, auth.as_ref());
-
-        let resp = match req.send().await {
-            Ok(r) => r,
-            Err(e) => {
-                debug!("Failed to reach upstream {url}: {e}");
-                return Vec::new();
-            }
-        };
-
-        if !resp.status().is_success() {
-            debug!(
-                "Upstream {url} returned {} for model discovery",
-                resp.status()
-            );
-            return Vec::new();
-        }
-
-        match resp.json::<Value>().await {
-            Ok(json) => ListModelsResponse::parse_upstream(&json, ProviderType::from_url(&url)),
-            Err(e) => {
-                warn!("Failed to parse upstream models from {url}: {e}");
-                Vec::new()
-            }
-        }
-    }
 }
 
 #[async_trait]
@@ -510,48 +385,6 @@ impl RouterTrait for RouterManager {
         } else {
             (StatusCode::SERVICE_UNAVAILABLE, "No routers available").into_response()
         }
-    }
-
-    async fn get_models(&self, req: Request<Body>) -> Response {
-        // Extract token from Authorization header (case-insensitive "Bearer " prefix
-        // per RFC 7235) or Anthropic-style x-api-key header.
-        let bearer_token = req
-            .headers()
-            .get(header::AUTHORIZATION)
-            .and_then(|h| h.to_str().ok())
-            .and_then(|h| {
-                let lower = h.to_ascii_lowercase();
-                lower.starts_with("bearer ").then(|| h[7..].to_string())
-            })
-            .or_else(|| {
-                req.headers()
-                    .get("x-api-key")
-                    .and_then(|h| h.to_str().ok())
-                    .map(String::from)
-            });
-
-        // Short-circuit: if the token matches any of the gateway's own
-        // credentials (shared or per-tenant), skip upstream fan-out and
-        // return registry models directly. A tenant-scoped key must never
-        // reach the BYOK fan-out below — that would forward it externally.
-        if let Some(ref token) = bearer_token {
-            if self.gateway_auth.contains_token(token) {
-                return self.registry_models_response();
-            }
-        }
-
-        // If the caller sent a provider token, try to discover models from
-        // upstream providers. This enables BYOK (bring your own key) flows.
-        if let Some(ref token) = bearer_token {
-            let upstream_cards = self.fetch_upstream_models(token).await;
-            if !upstream_cards.is_empty() {
-                let resp = ListModelsResponse::from_model_cards(upstream_cards);
-                return (StatusCode::OK, Json(resp)).into_response();
-            }
-            // All upstreams failed or returned nothing — fall through to registry.
-        }
-
-        self.registry_models_response()
     }
 
     async fn get_model_info(&self, req: Request<Body>) -> Response {
@@ -928,10 +761,10 @@ mod tests {
     use std::collections::HashMap;
 
     use async_trait::async_trait;
+    use openai_protocol::model_card::ModelCard;
 
     use super::*;
     use crate::{
-        config::TenantApiKeyEntry,
         middleware::{RouteRequestMeta, TenantKey},
         routers::factory::router_ids,
         worker::{BasicWorkerBuilder, CircuitBreakerConfig, WorkerRegistry},
@@ -1010,8 +843,7 @@ mod tests {
     }
 
     fn test_manager(enable_igw: bool) -> Arc<RouterManager> {
-        let mut manager =
-            RouterManager::new(Arc::new(WorkerRegistry::new()), reqwest::Client::new());
+        let mut manager = RouterManager::new(Arc::new(WorkerRegistry::new()));
         manager.enable_igw = enable_igw;
         let manager = Arc::new(manager);
         manager.register_router(router_ids::HTTP_REGULAR, Arc::new(StubRouter));
@@ -1042,7 +874,7 @@ mod tests {
         }
 
         let forward_manager = {
-            let mut m = RouterManager::new(Arc::new(WorkerRegistry::new()), reqwest::Client::new());
+            let mut m = RouterManager::new(Arc::new(WorkerRegistry::new()));
             m.enable_igw = false;
             let m = Arc::new(m);
             m.register_router(router_ids::HTTP_REGULAR, Arc::new(ForwardStubRouter));
@@ -1060,7 +892,7 @@ mod tests {
         );
 
         let pd_manager = {
-            let mut m = RouterManager::new(Arc::new(WorkerRegistry::new()), reqwest::Client::new());
+            let mut m = RouterManager::new(Arc::new(WorkerRegistry::new()));
             m.enable_igw = false;
             let m = Arc::new(m);
             m.register_router(router_ids::HTTP_PD, Arc::new(PdStubRouter));
@@ -1081,101 +913,6 @@ mod tests {
 
     fn test_tenant_meta() -> TenantRequestMeta {
         RouteRequestMeta::new(TenantKey::from("test-tenant"))
-    }
-
-    /// A tenant-scoped credential must never reach `fetch_upstream_models`:
-    /// that would forward the gateway's own secret to an external provider
-    /// as if it were the caller's BYOK token. Verified against a real mock
-    /// upstream that counts hits, since a stubbed HTTP client can't
-    /// distinguish "short-circuited" from "fell through and failed" by
-    /// response alone — both end up returning registry models on failure.
-    #[tokio::test]
-    #[expect(clippy::disallowed_methods, reason = "test infrastructure")]
-    async fn get_models_short_circuits_for_tenant_key_without_forwarding_upstream() {
-        use std::sync::atomic::{AtomicUsize, Ordering};
-
-        let hit_count = Arc::new(AtomicUsize::new(0));
-        let hit_count_clone = hit_count.clone();
-        let mock_app = axum::Router::new().route(
-            "/v1/models",
-            axum::routing::get(move || {
-                let hit_count = hit_count_clone.clone();
-                async move {
-                    hit_count.fetch_add(1, Ordering::SeqCst);
-                    Json(serde_json::json!({"object": "list", "data": []}))
-                }
-            }),
-        );
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-        tokio::spawn(async move {
-            let _ = axum::serve(listener, mock_app).await;
-        });
-
-        let registry = Arc::new(WorkerRegistry::new());
-        let external_worker: Arc<dyn Worker> = Arc::new(
-            BasicWorkerBuilder::new(format!("http://{addr}"))
-                .worker_type(WorkerType::Regular)
-                .runtime_type(RuntimeType::External)
-                .health_config(openai_protocol::worker::HealthCheckConfig {
-                    disable_health_check: true,
-                    ..Default::default()
-                })
-                .build(),
-        );
-        registry.register(external_worker);
-        // A non-External worker so registry_models_response() (the
-        // short-circuit path) has something to return besides 503.
-        let internal_worker: Arc<dyn Worker> = Arc::new(
-            BasicWorkerBuilder::new("http://internal.invalid:8000")
-                .worker_type(WorkerType::Regular)
-                .models(vec![ModelCard::new("internal-model")])
-                .health_config(openai_protocol::worker::HealthCheckConfig {
-                    disable_health_check: true,
-                    ..Default::default()
-                })
-                .build(),
-        );
-        registry.register(internal_worker);
-
-        let mut manager = RouterManager::new(registry, reqwest::Client::new());
-        manager.gateway_auth = AuthConfig::with_tenant_keys(
-            Some("shared-secret".to_string()),
-            &[TenantApiKeyEntry {
-                tenant_id: "team-red".to_string(),
-                key: "team-red-secret".to_string(),
-            }],
-        );
-
-        let req = Request::builder()
-            .method("GET")
-            .uri("/v1/models")
-            .header(header::AUTHORIZATION, "Bearer team-red-secret")
-            .body(Body::empty())
-            .unwrap();
-        let response = manager.get_models(req).await;
-        assert_eq!(response.status(), StatusCode::OK);
-        assert_eq!(
-            hit_count.load(Ordering::SeqCst),
-            0,
-            "a tenant-scoped key must never trigger upstream BYOK fan-out"
-        );
-
-        // Control: a genuinely unrecognized token should still trigger BYOK
-        // fan-out — confirms the short-circuit is keyed on the gateway's own
-        // credentials, not disabled entirely.
-        let req = Request::builder()
-            .method("GET")
-            .uri("/v1/models")
-            .header(header::AUTHORIZATION, "Bearer not-a-gateway-key")
-            .body(Body::empty())
-            .unwrap();
-        let _ = manager.get_models(req).await;
-        assert_eq!(
-            hit_count.load(Ordering::SeqCst),
-            1,
-            "an unrecognized token should still fan out to upstream providers"
-        );
     }
 
     fn generate_request_without_model() -> GenerateRequest {
@@ -1240,7 +977,7 @@ mod tests {
         add_workers(WorkerType::Decode, 2);
         add_workers(WorkerType::Regular, 6);
 
-        let mut manager = RouterManager::new(registry, reqwest::Client::new());
+        let mut manager = RouterManager::new(registry);
         manager.enable_igw = true;
         let manager = Arc::new(manager);
         manager.register_router(router_ids::HTTP_PD, Arc::new(PdStubRouter));
@@ -1284,7 +1021,7 @@ mod tests {
             registry.register(Arc::new(worker)).unwrap();
         }
 
-        let mut manager = RouterManager::new(registry, reqwest::Client::new());
+        let mut manager = RouterManager::new(registry);
         manager.enable_igw = true;
         let manager = Arc::new(manager);
         manager.register_router(router_ids::GRPC_PD, Arc::new(PdStubRouter));
@@ -1315,7 +1052,7 @@ mod tests {
             registry.register(Arc::new(worker)).unwrap();
         }
 
-        let mut manager = RouterManager::new(registry, reqwest::Client::new());
+        let mut manager = RouterManager::new(registry);
         manager.enable_igw = true;
         let manager = Arc::new(manager);
         manager.register_router(router_ids::HTTP_REGULAR, Arc::new(StubRouter));
